chore: add one-command local D1 seed script
