+++ packages/frontend-next/src/components/report/ReportSuccess.i18n.ts
@@ -5,6 +5,7 @@ export const NAMESPACE = 'reportSuccess';
 i18n.addResourceBundle('en', NAMESPACE, {
   title: 'Thank you for your report!',
   description: 'people appreciate your help',
+  feedbackHeading: 'Got a moment? Tell us what you think',
   syncText: 'Your report was synced with @FreiFahren_BE',
   continue: 'Continue',
 });
@@ -12,6 +13,7 @@ i18n.addResourceBundle('en', NAMESPACE, {
 i18n.addResourceBundle('de', NAMESPACE, {
   title: 'Danke für deine Meldung!',
   description: 'Menschen schätzen deine Hilfe',
+  feedbackHeading: 'Hast du kurz Zeit? Sag uns deine Meinung',
   syncText: 'Deine Meldung wurde mit @FreiFahren_BE synchronisiert.',
   continue: 'Weiter',
 });
