+++ packages/frontend-next/src/components/report/ReportSuccess.tsx
@@ -1,14 +1,16 @@
 import { Check, Users } from 'lucide-react';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import type { SubmitReportResponse } from '@/api/reports';
 import { useLines, useStations } from '@/api/transit';
-import { FeedbackButton } from '@/components/feedback/FeedbackButton';
+import { FeedbackForm } from '@/components/feedback/FeedbackForm';
 import { SocialLinks } from '@/components/map/SocialLinks';
 import { LineBadge } from '@/components/transit/LineBadge';
 import { Button } from '@/components/ui/button';
 import { useCountAnimation } from '@/hooks/useCountAnimation';
+import { captureSurveyShown } from '@/lib/analytics';
+import { FEEDBACK_SURVEY_ID } from '@/lib/feedback-modal';
 import { pickReporterCount } from '@/lib/reporters';
 
 import { NAMESPACE } from './ReportSuccess.i18n';
@@ -26,44 +28,63 @@ export function ReportSuccess({
   const [reporters] = useState(pickReporterCount);
   const count = useCountAnimation(reporters, 1500);
 
+  // The feedback form is embedded (not behind a button), so record that it was shown on arrival.
+  useEffect(() => {
+    captureSurveyShown(FEEDBACK_SURVEY_ID);
+  }, []);
+
   const stationName = stations?.[result.stationId]?.name;
   const lineName = result.lineId ? lines?.find((l) => l.id === result.lineId)?.name : null;
 
   return (
-    <div className="flex h-full flex-col items-center px-6 pt-16 pb-6 text-center">
-      <div className="bg-accent-bright text-primary-foreground animate-in zoom-in-50 fade-in flex size-16 items-center justify-center rounded-full shadow-[0_6px_16px_rgba(214,59,59,0.28)] duration-300">
-        <Check className="size-8" />
-      </div>
-      <h1 className="font-heading mt-5 text-2xl font-semibold">{t('title')}</h1>
-
-      {stationName && (
-        <div className="mt-4 flex items-center gap-2">
-          {lineName && <LineBadge name={lineName} />}
-          <span className="text-sm font-semibold">{stationName}</span>
+    <div className="flex h-full flex-col">
+      {/* Scrollable content: keeps the footer (Continue) pinned and always visible. */}
+      <div className="flex flex-1 flex-col items-center overflow-y-auto px-6 pt-6 pb-4 text-center">
+        <div className="flex items-center gap-3">
+          <div className="bg-accent-bright text-primary-foreground animate-in zoom-in-50 fade-in flex size-10 items-center justify-center rounded-full shadow-[0_6px_16px_rgba(214,59,59,0.28)] duration-300">
+            <Check className="size-6" />
+          </div>
+          <h1 className="font-heading text-xl font-semibold">{t('title')}</h1>
         </div>
-      )}
 
-      <div className="flex flex-1 flex-col items-center justify-center gap-1">
-        <div className="text-foreground flex items-center gap-2">
-          <Users className="text-muted-foreground size-6" />
-          <span className="font-heading text-4xl font-bold tabular-nums">
-            {count.toLocaleString()}
-          </span>
+        {stationName && (
+          <div className="mt-4 flex items-center gap-2">
+            {lineName && <LineBadge name={lineName} />}
+            <span className="text-sm font-semibold">{stationName}</span>
+          </div>
+        )}
+
+        <div className="mt-5 flex flex-col items-center gap-1">
+          <div className="text-foreground flex items-center gap-2">
+            <Users className="text-muted-foreground size-6" />
+            <span className="font-heading text-4xl font-bold tabular-nums">
+              {count.toLocaleString()}
+            </span>
+          </div>
+          <p className="text-muted-foreground text-sm">{t('description')}</p>
         </div>
-        <p className="text-muted-foreground text-sm">{t('description')}</p>
-        <FeedbackButton source="report_success" variant="outline" className="border-border mt-4" />
+
+        <div className="mt-6 w-full text-left">
+          <h2 className="text-muted-foreground mb-2 text-center text-sm font-semibold">
+            {t('feedbackHeading')}
+          </h2>
+          <FeedbackForm compact />
+        </div>
       </div>
 
-      <p className="text-muted-foreground text-[0.6875rem]">{t('syncText')}</p>
-      <Button
-        type="button"
-        size="lg"
-        onClick={onClose}
-        className="bg-accent-bright text-primary-foreground hover:bg-accent-press mt-4 h-12 w-full rounded-lg text-base font-semibold shadow-[0_6px_16px_rgba(214,59,59,0.28)]"
-      >
-        {t('continue')}
-      </Button>
-      <SocialLinks appearance="inline" className="mt-3 justify-center" />
+      {/* Pinned footer — the primary action stays reachable regardless of form size. */}
+      <div className="border-border/60 bg-card flex shrink-0 flex-col items-center border-t px-6 pt-3 pb-6">
+        <p className="text-muted-foreground text-[0.6875rem]">{t('syncText')}</p>
+        <Button
+          type="button"
+          size="lg"
+          onClick={onClose}
+          className="bg-accent-bright text-primary-foreground hover:bg-accent-press mt-2 h-12 w-full rounded-lg text-base font-semibold shadow-[0_6px_16px_rgba(214,59,59,0.28)]"
+        >
+          {t('continue')}
+        </Button>
+        <SocialLinks appearance="inline" className="mt-3 justify-center" />
+      </div>
     </div>
   );
 }
